refactor(stream): extract LiveSetupTest from FlowLive
Device selection, level meter and the /test rehearsal broadcast move into a
shared component so the show dashboard can render live preparation inline.
FlowLive becomes a thin wrapper (step chrome + navigation only); the
useAudioCapture/useStreamSocket singletons keep behaviour identical.
